Refactor app
