+++ api/helpers/auth.py
@@ -39,13 +39,10 @@ def get_current_user():
     """Fetches current user details from table"""
     token = request.headers['token']
     decoded_token = decode_token(token)
-    # try:
     userId = decoded_token["userId"]
     for user_obj in users_table:
         if user_obj.userId == userId:
             return {"userId": userId, "isAdmin": user_obj.isAdmin}
-    # except KeyError:
-    #     return jsonify({"message": "userId not in token"}), 401
 
 
 
@@ -69,8 +66,3 @@ def non_admin_required(func):
         return func(*args, **kwargs)
     return wrapper
 
-
-def encode_token_test(userId):
-    token = jwt.encode({'userId': userId, 'exp': datetime.datetime.utcnow() + datetime.timedelta(hours=20)},
-        "secret_key")
-    return token
