+++ tests/test_incident.py
@@ -5,23 +5,18 @@ import sys
 
 sys.path.append(os.path.dirname(os.path.abspath(__file__)) + "/..")
 
-from .test_base import new_intervention,error,new_location,new_status,\
+from .test_base import new_intervention,new_intervention1,error,new_location,new_status,\
                         new_bad_intervention,new_error_intervention,\
-                        new_comment,new_error_redflag,\
-                        new_intervention_response,new_redflag,\
-                        new_redflag_response,token_header
+                        new_comment,new_error_redflag,new_redflag1,\
+                        new_intervention_response,new_intervention1_response,new_redflag,\
+                        new_redflag_response,new_redflag1_response,token_header
 from api.app import app
 from api.helpers.auth import encode_token
 from api.helpers.incidenthelper import get_incidents_by_type
+from .test_base import TestBase
 
 
-
-class TestIntervention(unittest.TestCase):
-    def setUp(self):
-        app.config['TESTING'] = True
-        app.config['DEBUG'] = True
-        self.app = app.test_client()
-        self.assertFalse(app.config['SECRET_KEY'] is 'softwareDeveloper.Manuel@secret_key/mats.com')
+class TestIntervention(TestBase):
 
     def test_index(self):
         response = self.app.get('/api/v1/')
@@ -30,6 +25,39 @@ class TestIntervention(unittest.TestCase):
         message = {
             'IReporter': "This enables any/every citizen to bring any form of corruption to the notice of appropriate authorities and the general public."}
         self.assertEqual(json.loads(data), message)
+
+
+    def test_create_redflag(self):
+        response = self.app.post('/api/v1/red-flags', headers=token_header(encode_token(2)),
+                                     data=json.dumps(new_redflag))
+        self.assertEqual(response.status_code,201)
+        data = response.data.decode()
+        self.assertEqual(json.loads(data), new_redflag_response)
+
+
+    def test_create_intervention(self):
+        response = self.app.post('/api/v1/intervention', headers=token_header(encode_token(2)),
+                                 data=json.dumps(new_intervention))
+        self.assertEqual(response.status_code,201)
+        data = response.data.decode()
+        self.assertEqual(json.loads(data), new_intervention_response)
+
+    def test_create_more_redflag(self):
+        response = self.app.post('/api/v1/red-flags', headers=token_header(encode_token(2)),
+                                     data=json.dumps(new_redflag1))
+        self.assertEqual(response.status_code,201)
+        data = response.data.decode()
+        self.assertEqual(json.loads(data), new_redflag1_response)
+
+
+    def test_create_more_intervention(self):
+        response = self.app.post('/api/v1/intervention', headers=token_header(encode_token(2)),
+                                 data=json.dumps(new_intervention1))
+        self.assertEqual(response.status_code,201)
+        data = response.data.decode()
+        self.assertEqual(json.loads(data), new_intervention1_response)
+
+
 
     def test_get_all_redflags(self):
         response = self.app.get('/api/v1/red-flags', headers=token_header(encode_token(2)))
@@ -63,7 +91,11 @@ class TestIntervention(unittest.TestCase):
                 },
                 "status": 200
             }
-        self.assertEqual(json.loads(data), message)
+        self.assertEqual(json.loads(data)["data"]["comment"], message["data"]["comment"])
+        self.assertEqual(json.loads(data)["data"]["createdBy"], message["data"]["createdBy"])
+        self.assertEqual(json.loads(data)["data"]["locationLong"], message["data"]["locationLong"])
+        self.assertEqual(json.loads(data)["data"]["locationLat"], message["data"]["locationLat"])
+        self.assertEqual(json.loads(data)["data"]["type"], message["data"]["type"])
 
 
     def test_get_specific_redflags(self):
@@ -87,7 +119,12 @@ class TestIntervention(unittest.TestCase):
                     "status": 200
                 }
             ]
-        self.assertEqual(json.loads(data), message)
+
+        self.assertEqual(json.loads(data)[0]["data"]["comment"], message[0]["data"]["comment"])
+        self.assertEqual(json.loads(data)[0]["data"]["createdBy"], message[0]["data"]["createdBy"])
+        self.assertEqual(json.loads(data)[0]["data"]["locationLong"], message[0]["data"]["locationLong"])
+        self.assertEqual(json.loads(data)[0]["data"]["locationLat"], message[0]["data"]["locationLat"])
+        self.assertEqual(json.loads(data)[0]["data"]["type"], message[0]["data"]["type"])
 
 
     def test_update_intervention_location(self):
@@ -130,54 +167,39 @@ class TestIntervention(unittest.TestCase):
 
  
     def test_delete_intervention(self):
-        response = self.app.delete('/api/v1/intervention/3', headers=token_header(encode_token(2)))
+        response = self.app.delete('/api/v1/intervention/1', headers=token_header(encode_token(2)))
         self.assertEqual(response.status_code,200)
         data = response.data.decode()
-        message = {"data": {"id": 3, "message": "intervention record has been deleted"},
+        message = {"data": {"id": 1, "message": "intervention record has been deleted"},
                    "status": 200}
         self.assertEqual(json.loads(data), message)
 
 
     def test_delete_redflag(self):
-        response = self.app.delete('/api/v1/red-flags/3', headers=token_header(encode_token(2)))
+        response = self.app.delete('/api/v1/red-flags/1', headers=token_header(encode_token(2)))
         self.assertEqual(response.status_code,200)
         data = response.data.decode()
-        message = {"data": {"id": 3, "message": "redflag record has been deleted"},
+        message = {"data": {"id": 1, "message": "redflag record has been deleted"},
                    "status": 200}
         self.assertEqual(json.loads(data), message)
 
-
-    def test_create_redflag(self):
-        response = self.app.post('/api/v1/red-flags', headers=token_header(encode_token(2)), data=json.dumps(new_redflag))
-        self.assertEqual(response.status_code,201)
-        data = response.data.decode()
-        self.assertEqual(json.loads(data), new_redflag_response)
-
-
-    def test_create_intervention(self):
-        response = self.app.post('/api/v1/intervention', headers=token_header(encode_token(2)),
-                                 data=json.dumps(new_intervention))
-        self.assertEqual(response.status_code,201)
-        data = response.data.decode()
-        self.assertEqual(json.loads(data), new_intervention_response)
-
  
     def test_update_intervention_status(self):
-        response = self.app.patch('/api/v1/intervention/1/status', headers=token_header(encode_token(1)),
+        response = self.app.patch('/api/v1/intervention/2/status', headers=token_header(encode_token(1)),
                                   data=json.dumps(new_status))
         self.assertEqual(response.status_code,200)
         data = response.data.decode()
-        message = {"data": [{"id": 1, "message": "Updated intervention record's status"}],
+        message = {"data": [{"id": 2, "message": "Updated intervention record's status"}],
                    "status": 200}
         self.assertEqual(json.loads(data), message)
 
 
     def test_update_redflag_status(self):
-        response = self.app.patch('/api/v1/red-flags/1/status', headers=token_header(encode_token(1)),\
+        response = self.app.patch('/api/v1/red-flags/2/status', headers=token_header(encode_token(1)),\
                                           data=json.dumps(new_status))
         self.assertEqual(response.status_code,200)
         data = response.data.decode()
-        message = {"data": {"id": 1, "message": "Updated redflag record's status"},
+        message = {"data": {"id": 2, "message": "Updated redflag record's status"},
                    "status": 200}
         self.assertEqual(json.loads(data), message)
