+++ tests/test_incident_error.py
@@ -12,16 +12,12 @@ from .test_base import new_intervention,invalid_key_msg,error,example_create_dat
 from api.app import app
 from api.helpers.auth import encode_token
 from api.helpers.incidenthelper import get_incidents_by_type
+from .test_base import TestBase
 
 
 
-class TestIntervention(unittest.TestCase):
-    def setUp(self):
-        app.config['TESTING'] = True
-        app.config['DEBUG'] = True
-        self.app = app.test_client()
-        self.assertFalse(app.config['SECRET_KEY'] is 'softwareDeveloper.Manuel@secret_key/mats.com')
-
+class TestIntervention(TestBase):
+    
     def test_index(self):
         response = self.app.get('/api/v1/')
         self.assertEqual(response.status_code,200)
@@ -159,7 +155,7 @@ class TestIntervention(unittest.TestCase):
         self.assertEqual(json.loads(data), error)
 
     def test_update_not_possible_comment(self):
-        response = self.app.patch('/api/v1/intervention/1/comment', headers=token_header(encode_token(2)),
+        response = self.app.patch('/api/v1/intervention/2/comment', headers=token_header(encode_token(2)),
                                   data=json.dumps(new_comment))
         self.assertEqual(response.status_code,406)
         data = response.data.decode()
